Extract arithmetic operator utils to a separate file
